Wait for cloud init to finish
We have race conditions that all stem from the fact that terraform
doesn't wait for cloud-init to finish. In most cases, things work out
eventually, but it's most prudent to just wait it out before we
jumpstart the k0s init process. In order to do that, we need a signal
from the server that the process is done. The easiest way to do so, is
to use the remote-exec provision and call cloud-init status --wait. This
is a PoC for that

Signed-off-by: Alexandros Kosiaris <akosiaris@gmail.com>