Move k0s, helm into own module
Create a new module out of helm and k0s resource files. Move the needed
template in there, add new variables and call the module from the root
module. The k0s module also needs to declare the providers it is going
to use. Interestingly, it's the only module that uses other providers
too, the root module isn't requiring the hcloud one apparently

Signed-off-by: Alexandros Kosiaris <akosiaris@gmail.com>