Obtain and use the IP addresses from the server module
hcloud_primary_ip returns a /64 for IPv6 which isn't directly usable.
While we could work around it, hcloud_server actually returns the IP
used by the server itself.

Expose the IPs of the servers from the server module and use them in the
place of the IPs from the network module. This also does away with the
need to have k0s explicitly depend on the server module calls, so remove
that too

Signed-off-by: Alexandros Kosiaris <akosiaris@gmail.com>