server: Add Hetzner firewalling capability
Add firewalling capability to the server child module. Utilize it in the
case of controllers to offer the exact same functionality as ferm did,
without the downside of messing with terraform state when adding nodes

Workers are left without firewalling functionality in this patch on
purpose. A followup patch implementing Calico HostEndpoints will take
care of that

Signed-off-by: Alexandros Kosiaris <akosiaris@gmail.com>